+++ Plugin/Thunderstore/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fixed Whistle and plushie sounds being picked up/dropped being global.
 - Allowed Redwood Giant to target old birds and Driftwood Giants.
 - Fixed bug with Redwood Giant being stuck after eating a corpse of a ForestGiant.
+- Added death animation for Redwood Giant.
 
 <details>
   <summary>Older Versions</summary>